Relocated css, js, and images to respective folders. Also updated App.js file
