+++ evstreet/src/components/App.js
@@ -1,0 +1,15 @@
+import logo from "../assets/images/logo.svg";
+import "../assets/styles/App.css";
+
+function App() {
+  return (
+    <div className="App">
+      <header className="App-header">
+        <img src={logo} className="App-logo" alt="logo" />
+      </header>
+      <h1>Hello</h1>
+    </div>
+  );
+}
+
+export default App;
